# 03_Backend/Innovatube/Innovatube/Models/Login.cs
using System.ComponentModel.DataAnnotations;

namespace Innovatube.Models
{
    public class Login
    {
        public string? login {  get; set; }
        public string? accessPassword { get; set; }
    }
}