# 03_Backend/Innovatube/Innovatube/Models/Video.cs
namespace Innovatube.Models
{
    public class Video
    {
        public string? IdVideo { get; set; }
        public string? Title { get; set; }
        public string? ThumbnailUrl { get; set; }
        public DateTime? PublishedAt { get; set; }
    }
}