# 03_Backend/Innovatube/Innovatube/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Innovatube",
    "version": "v1"
  },
  "paths": {
    "/api/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/forgot-password": {
      "post": {
        "tags": [
          "Password"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ForgotPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/reset-password": {
      "post": {
        "tags": [
          "Password"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ResetPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RegisterUsers": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/videos/favorites": {
      "post": {
        "tags": [
          "Videos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Video"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/videos/favorites/{userId}": {
      "get": {
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/videos/favorites/search": {
      "get": {
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/videos/favorites/{userId}/{videoId}": {
      "delete": {
        "tags": [
          "Videos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "videoId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "ForgotPasswordRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "Login": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string"
        },
        "accessPassword": {
          "type": "string"
        }
      }
    },
    "RegisterUserRequest": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "accessPassword": {
          "type": "string"
        },
        "recaptchaToken": {
          "type": "string"
        }
      }
    },
    "ResetPasswordRequest": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      }
    },
    "Video": {
      "type": "object",
      "properties": {
        "idVideo": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "thumbnailUrl": {
          "type": "string"
        },
        "publishedAt": {
          "format": "date-time",
          "type": "string"
        }
      }
    }
  }
}